+++ react/components/SKUSelector/components/SKUSelector.tsx
@@ -121,6 +121,35 @@ interface AvailableVariationParams {
   sortVariationsByLabel: boolean
 }
 
+export const orderItemsByAvailability = (
+  item1: SelectorProductItem,
+  item2: SelectorProductItem
+) => {
+  const isAvailableItem1 = isSkuAvailable(item1)
+  const isAvailableItem2 = isSkuAvailable(item2)
+
+  if (isAvailableItem1 && !isAvailableItem2) {
+    return -1
+  }
+
+  if (!isAvailableItem1 && isAvailableItem2) {
+    return 1
+  }
+
+  const defaultSellerItem1 = getDefaultSeller(item1.sellers)
+  const defaultSellerItem2 = getDefaultSeller(item2.sellers)
+
+  const availabilityItem1 = defaultSellerItem1
+    ? defaultSellerItem1.commertialOffer.AvailableQuantity
+    : 0
+
+  const availabilityItem2 = defaultSellerItem2
+    ? defaultSellerItem2.commertialOffer.AvailableQuantity
+    : 0
+
+  return availabilityItem2 - availabilityItem1
+}
+
 const parseOptionNameToDisplayOption =
   ({
     selectedVariations,
@@ -161,13 +190,20 @@ const parseOptionNameToDisplayOption =
 
     if (possibleItems.length > 0) {
       // This is a valid combination option
-      const [item] = possibleItems
+
+      const possibleItemsOrderedByAvailability =
+        possibleItems.length > 1
+          ? possibleItems.sort(orderItemsByAvailability)
+          : possibleItems
+
+      const [item] = possibleItemsOrderedByAvailability
+
       const callbackFn = onSelectItemMemo({
         name: variationName,
         value: variationValue.name,
         skuId: item.itemId,
         isMainAndImpossible: false,
-        possibleItems,
+        possibleItems: possibleItemsOrderedByAvailability,
       })
 
       return {
@@ -176,7 +212,7 @@ const parseOptionNameToDisplayOption =
         onSelectItem: callbackFn,
         image,
         available: showItemAsAvailable({
-          possibleItems,
+          possibleItems: possibleItemsOrderedByAvailability,
           selectedVariations,
           variationCount,
           isSelected,
@@ -248,13 +284,16 @@ const variationNameToDisplayVariation =
       const allNumbers = options.every(
         (option: any) => !Number.isNaN(option.label)
       )
+
       options.sort((a: any, b: any) => {
         if (allNumbers) {
           return a.label - b.label
         }
+
         return a.label < b.label ? -1 : a.label > b.label ? 1 : 0
       })
     }
+
     return { name, originalName, options }
   }