+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Default selection logic in `SKU Selector` component to prioritize available SKUs.
+
 ## [3.175.1] - 2024-09-11
 
 ### Fixed